encapsulate autodiscovery utils in a separate file and remove autodiscovery for agentpool mode
